@@ -1,10 +1,8 @@
-// import * as ts from "typescript"
-import TS = require("typescript");
+import * as ts from "typescript"
 import { IMark } from "../type";
 import { createMark, isValidNpmPackageName } from "../utils";
 
 export function compile(code: string, filepath: string): IMark[] {
-  const ts: typeof TS = require("typescript");
   const marks: IMark[] = [];
   let sourceFile;
   try {
@@ -21,23 +19,23 @@ export function compile(code: string, filepath: string): IMark[] {
     return [];
   }
 
-  function delint(SourceFile: TS.SourceFile) {
+  function delint(SourceFile: ts.SourceFile) {
     delintNode(SourceFile);
 
-    function delintNode(node: TS.Node) {
-      let moduleNode: TS.LiteralLikeNode | null = null;
+    function delintNode(node: ts.Node) {
+      let moduleNode: ts.LiteralLikeNode | null = null;
       switch (node.kind) {
         // require('xxx')
         // import('xxx')
         case ts.SyntaxKind.CallExpression:
-          const expression = (node as TS.CallExpression).expression;
-          const args = (node as TS.CallExpression).arguments;
+          const expression = (node as ts.CallExpression).expression;
+          const args = (node as ts.CallExpression).arguments;
           const isRequire =
             ts.isIdentifier(expression) && expression.text === "require";
           const isDynamicImport =
             expression.kind === ts.SyntaxKind.ImportKeyword;
           if (isRequire || isDynamicImport) {
-            const argv = args[0] as TS.StringLiteral;
+            const argv = args[0] as ts.StringLiteral;
 
             if (argv && ts.isStringLiteral(argv)) {
               moduleNode = argv;
@@ -46,8 +44,8 @@ export function compile(code: string, filepath: string): IMark[] {
           break;
         // import ts = require('typescript')
         case ts.SyntaxKind.ImportEqualsDeclaration:
-          const ref = (node as TS.ImportEqualsDeclaration)
-            .moduleReference as TS.ExternalModuleReference;
+          const ref = (node as ts.ImportEqualsDeclaration)
+            .moduleReference as ts.ExternalModuleReference;
 
           if (ts.isStringLiteral(ref.expression)) {
             moduleNode = ref.expression;
@@ -57,15 +55,15 @@ export function compile(code: string, filepath: string): IMark[] {
         // import 'xx'
         // import xx from 'xx'
         case ts.SyntaxKind.ImportDeclaration:
-          const spec = (node as TS.ImportDeclaration).moduleSpecifier;
+          const spec = (node as ts.ImportDeclaration).moduleSpecifier;
           if (spec && ts.isStringLiteral(spec)) {
             moduleNode = spec;
           }
           break;
-        // export { window } from "vscode";
-        // export * from "vscode";
+        // export { window } from "xxx";
+        // export * from "xxx";
         case ts.SyntaxKind.ExportDeclaration:
-          const exportSpec = (node as TS.ExportDeclaration).moduleSpecifier;
+          const exportSpec = (node as ts.ExportDeclaration).moduleSpecifier;
           if (exportSpec && ts.isStringLiteral(exportSpec)) {
             moduleNode = exportSpec;
           }
